test run on a pc

+++ twskbino/fifo_queue.cpp
@@ -20,7 +20,9 @@ KeyFifo::key_fifo_data_t *KeyFifo::popkd(unsigned long ctsms, int gapms)
 {
 	key_fifo_data_t *kd;
 	if(outp==inp){return NULL;}
-	if((ctsms - kds[(inp-1)&FIFO_MAX_DEPTH_MASK].tsms) < gapms){return NULL;}
+	if((ctsms - kds[(inp-1)&FIFO_MAX_DEPTH_MASK].tsms) < (unsigned long)gapms){
+		return NULL;
+	}
 	kd=&kds[outp++];
 	outp&=FIFO_MAX_DEPTH_MASK;
 	return kd;
